refactor(identifiers): promote sanitize_userinfo to public utility + warn on redaction
Milestone 105 phase 2A. Promotes the milestone-075 URL credential
redaction helper from module-private state in
binding/identifiers/auto_detect.rs to a shared public utility at
mikebom-cli/src/identifiers/sanitize.rs so the new C/C++ readers
(west.yml, idf_component.yml, git-submodule, plus cmake/conan
extensions) can call it for FR-016 unconditional URL sanitization.

API simplification: the previous `SanitizedUrl { original, sanitized,
was_sanitized }` struct collapses to `pub fn sanitize_userinfo(&str)
-> Cow<'_, str>`. Callers derive "was_sanitized" from
`matches!(result, Cow::Owned(_))`. Behavioral refinement: empty-userinfo
inputs (`https://@host/...`) now return Cow::Borrowed since there's
no real credential to strip — Cow::Owned cleanly indicates "real
credentials were stripped" for FR-016 log gating.

Log level bumped from tracing::info! to tracing::warn! at the three
existing call sites (per FR-016). The helper itself stays log-free
so per-callsite context (scheme=repo / scheme=git) is preserved.

Tests: 12 unit tests moved from auto_detect.rs's test module to the
new sanitize.rs test module. Net diff is -322 / +46 LOC + the new
~270-line sanitize.rs file. Pre-PR gate passes clean.

Constitution: Principle VI honored (no new crate; lib-side module
exposure follows the established `binding` pattern). Principle V
unaffected (this is internal-helper refactoring; no SBOM output
shape change).
